move chart wrapper
