+++ src/containers/BugzillaGraph/index.jsx
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { withErrorBoundary } from '../../vendor/errors';
-import ChartJsWrapper from '../../components/ChartJsWrapper';
+import ChartJsWrapper from '../../vendor/chartJs/ChartJsWrapper';
 import getBugsData from '../../utils/bugzilla/getBugsData';
 
 class BugzillaGraph extends Component {
