+++ src/containers/PerfherderGraphContainer/index.jsx
@@ -2,8 +2,8 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import LinkIcon from '@material-ui/icons/Link';
 import { withStyles } from '@material-ui/core/styles';
-import ChartJsWrapper from '../../components/ChartJsWrapper';
-import CustomTooltip from '../../utils/chartJs/CustomTooltip';
+import ChartJsWrapper from '../../vendor/chartJs/ChartJsWrapper';
+import CustomTooltip from '../../vendor/chartJs/CustomTooltip';
 import { withErrorBoundary } from '../../vendor/errors';
 import { exists, missing } from '../../vendor/utils';
 import { toQueryString } from '../../vendor/convert';
@@ -11,7 +11,7 @@ import Date from '../../vendor/dates';
 import { getData, TREEHERDER } from '../../vendor/perfherder';
 import { selectFrom } from '../../vendor/vectors';
 import { Log } from '../../vendor/logs';
-import generateDatasetStyle from '../../utils/chartJs/generateDatasetStyle';
+import generateDatasetStyle from '../../vendor/chartJs/generateDatasetStyle';
 import SETTINGS from '../../settings';
 
 const DEFAULT_TIMERANGE = Date.newInstance('today-6week').unix();
