+++ src/nimbledroid/NimbledroidGraphContainer.jsx
@@ -2,9 +2,9 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { CONFIG } from './config';
 import { withErrorBoundary } from '../vendor/errors';
-import ChartJsWrapper from '../components/ChartJsWrapper';
+import ChartJsWrapper from '../vendor/chartJs/ChartJsWrapper';
 import fetchNimbledroidData from './fetchNimbledroidData';
-import generateDatasetStyle from '../utils/chartJs/generateDatasetStyle';
+import generateDatasetStyle from '../vendor/chartJs/generateDatasetStyle';
 import SETTINGS from '../settings';
 import { selectFrom, toPairs } from '../vendor/vectors';
 import Date from '../vendor/dates';
